Fixed bug with save Food

+++ src/main/java/com/ordering_system/service/impl/FoodServiceImpl.java
@@ -3,17 +3,22 @@ package com.ordering_system.service.impl;
 
 import com.ordering_system.model.domain.FoodEntity;
 import com.ordering_system.model.domain.RestaurantEntity;
+import com.ordering_system.model.domain.UserEntity;
 import com.ordering_system.model.dto.Food;
+import com.ordering_system.model.enumeration.Role;
 import com.ordering_system.model.exception.ActivationException;
 import com.ordering_system.model.exception.EntityAlreadyExistsException;
 import com.ordering_system.repository.FoodRepository;
 import com.ordering_system.repository.RestaurantRepository;
+import com.ordering_system.repository.UserRepository;
 import com.ordering_system.service.FoodService;
 import com.ordering_system.service.converter.Converter;
+import com.ordering_system.service.mailsender.GetMail;
 import com.ordering_system.service.validator.Validator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,15 +29,18 @@ public class FoodServiceImpl implements FoodService {
     private final FoodRepository foodRepository;
     private final Converter converter;
     private final RestaurantRepository restaurantRepository;
-
+    private final Validator validator;
     private static final Logger LOGGER = LoggerFactory.getLogger(FoodServiceImpl.class);
 
     @Autowired
     public FoodServiceImpl(FoodRepository foodRepository, Converter converter,
-                           RestaurantRepository restaurantRepository) {
+                           RestaurantRepository restaurantRepository,
+                           UserRepository userRepository,
+                           Validator validator) {
         this.foodRepository = foodRepository;
         this.converter = converter;
         this.restaurantRepository = restaurantRepository;
+        this.validator = validator;
     }
 
     @Override
@@ -56,9 +64,10 @@ public class FoodServiceImpl implements FoodService {
         Validator.checkId(food.getRestaurantId());
         RestaurantEntity restaurantEntity = restaurantRepository.findRestaurantEntityById(food.getRestaurantId());
         Validator.checkEntity(restaurantEntity);
-        
+   
+        validator.checkAccess(restaurantEntity);
         if(restaurantEntity.isActivated() == false) {
-        	throw new ActivationException("Restaurant Entity isn't activated");
+        	throw new ActivationException("Restaurant isn't activated");
         }
         if(foodRepository.findByRestaurantEntityAndName(restaurantEntity, food.getName()) != null) {
         	throw new EntityAlreadyExistsException("Food by entered name already exists");
@@ -72,7 +81,9 @@ public class FoodServiceImpl implements FoodService {
     public void update(long id, Food food) {
         LOGGER.info("In method update in FoodServiceImpl class");
         FoodEntity foodEntity = foodRepository.findFoodEntityById(id);
+        RestaurantEntity restaurantEntity = restaurantRepository.findRestaurantEntityById(food.getRestaurantId());
 
+        validator.checkAccess(restaurantEntity);
         Validator.checkEntity(food);
         if (food.getName() != null) {
             Validator.checkName(food.getName());
@@ -95,8 +106,10 @@ public class FoodServiceImpl implements FoodService {
     @Override
     public void delete(long id) {
         LOGGER.info("In method delete in FoodServiceImpl class");
+        FoodEntity foodEntity = foodRepository.findFoodEntityById(id);
         Validator.checkId(id);
-        Validator.checkEntity(foodRepository.findFoodEntityById(id));
+        Validator.checkEntity(foodEntity);
+        validator.checkAccess(foodEntity.getRestaurantEntity());
         foodRepository.deleteById(id);
         LOGGER.info("Delete method passed in FoodServiceImpl class");
     }
